♻️ [REFACTOR]: 공통 부분인 MessageCard Header부분 컴포넌트 분리
