+++ src/components/CardBadage/CardBadage.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { Badge } from './CardBadage.style';
+
+export const CardBadage = ({ $type }) => {
+  return <Badge $type={$type}>{$type}</Badge>;
+};
