+++ src/components/MessageCard/MessageCard.style.js
@@ -19,48 +19,8 @@ export const Wrapper = styled.div`
   }
 `;
 
-export const TopWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  border-bottom: 1px solid #eeeeee;
-  padding: 0 0 15px;
-  justify-content: space-between;
-`;
-
-export const Image = styled.img`
-  width: 56px;
-  height: 56px;
-  border-radius: 100px;
-  border: 1px solid #ffffff;
-`;
-
-export const ProfileWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-`;
-export const ProfileTextWrapper = styled.div`
-  display: flex;
-  gap: 6px;
-`;
-
-const ProfileText = styled.p`
-  font-size: 20px;
-  color: #000000;
-  line-height: 24px;
-`;
-
-export const ProfileTextHead = styled(ProfileText)`
-  font-weight: 400;
-`;
-
-export const ProfileName = styled(ProfileText)`
-  font-weight: 700;
-`;
-
 export const TextWrapper = styled.div`
   margin-top: 15px;
-  text-overflow: ellipsis;
   overflow: hidden;
   font-size: 18px;
   font-weight: 400;
@@ -69,7 +29,6 @@ export const TextWrapper = styled.div`
   display: -webkit-box;
   -webkit-line-clamp: 3;
   -webkit-box-orient: vertical;
-  word-wrap: break-word;
 `;
 
 export const CreatedDate = styled.div`
@@ -82,6 +41,7 @@ export const CreatedDate = styled.div`
   font-size: 12px;
   line-height: 18px;
   color: #999999;
+  background-color: #ffffff;
 `;
 
 export const DeleteButton = styled.button`
