+++ src/components/Modal/Modal.jsx
@@ -1,21 +1,12 @@
 import React, { useContext } from 'react';
 import {
   Wrapper,
-  TopWrapper,
-  ProfileModalWrapper,
   CreatedDate,
   TextWrapper,
   ModalButton,
   Text,
 } from './Modal.style';
-import {
-  Image,
-  ProfileWrapper,
-  ProfileTextWrapper,
-  ProfileTextHead,
-  ProfileName,
-} from '../MessageCard/MessageCard.style';
-import { MessageCardBadge } from '../../pages/PostIDPage.style';
+import { MessageCardHeader } from '../MessageCardHeader/MessageCardHeader';
 import { PostIDContext } from '../../context/PostIDContext';
 
 export const Modal = () => {
@@ -33,30 +24,17 @@ export const Modal = () => {
   const handleModalButton = () => {
     handleCurrentCardData();
   };
+  const CreatedDateComponent = () => {
+    return <CreatedDate>{formattedDate}</CreatedDate>;
+  };
   return (
     <>
       {cardData.id && (
         <Wrapper>
-          <TopWrapper>
-            <ProfileModalWrapper>
-              <Image
-                src={cardData.profileImageURL}
-                alt="프로필 이미지"
-                width={56}
-                height={56}
-              ></Image>
-              <ProfileWrapper>
-                <ProfileTextWrapper>
-                  <ProfileTextHead>From.</ProfileTextHead>
-                  <ProfileName>{cardData.sender}</ProfileName>
-                </ProfileTextWrapper>
-                <MessageCardBadge $type={cardData.relationship}>
-                  {cardData.relationship}
-                </MessageCardBadge>
-              </ProfileWrapper>
-            </ProfileModalWrapper>
-            <CreatedDate>{formattedDate}</CreatedDate>
-          </TopWrapper>
+          <MessageCardHeader
+            cardData={cardData}
+            Component={CreatedDateComponent}
+          ></MessageCardHeader>
           <TextWrapper>
             <Text>{cardData.content}</Text>
           </TextWrapper>
