+++ src/style/GlobalStyle.js
@@ -4,7 +4,7 @@ import Pretendard from '../assets/fonts/Pretendard-Regular.otf';
 
 const GlobalStyle = createGlobalStyle`
   @font-face {
-    font-family: 'Pretendard-Regular';
+    font-family: 'Pretendard';
     font-weight: normal;
     src: url(${Pretendard}) format('truetype');
   }
